+++ src/app/(home)/components/index.ts
@@ -1,3 +1,4 @@
 export { default as Header } from './Header';
 export { default as HomeBottomSheet } from './HomeBottomSheet';
 export { default as MapSection } from './MapSection';
+export * from './SheetProvider';
